requirement.txt
django
django-crispy-forms
django-filter
djangorestframework
Pillow
